pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentpk"
version = "0.2.0"
description = "Package AI agents into portable .agent files"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Nomotic AI" },
]
keywords = ["ai", "agents", "packaging", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "cryptography>=41.0",
]

[project.scripts]
agent = "agentpk.cli:cli"

[project.urls]
Homepage = "https://github.com/nomoticai/agentpk"
Repository = "https://github.com/nomoticai/agentpk"

[tool.hatch.build.targets.wheel]
packages = ["src/agentpk"]

[tool.setuptools.package-data]
agentpk = [
    "extractors/js_ast_helper.js",
    "api/ui/*.html",
    "api/ui/*.css",
    "api/ui/*.js",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
]
api = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",
]

